requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.4
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
python-multipart==0.0.12
PyJWT==2.9.0
opencv-python==4.10.0.84
mediapipe==0.10.14
numpy==2.1.3
face-recognition==1.3.0
pgvector==0.3.6
httpx==0.27.2
